feat(settings): improve map pin ux
Closes #2999
