+++ packages/components/src/MapWithPin/MapPin.stories.tsx
@@ -1,0 +1,21 @@
+import { MapPin } from './MapPin'
+
+import type { Meta, StoryFn } from '@storybook/react'
+
+export default {
+  title: 'Components/MapPin',
+  component: MapPin,
+} as Meta<typeof MapPin>
+
+export const Default: StoryFn<typeof MapPin> = () => {
+  const position = { lat: 0, lng: 0 }
+  return (
+    <MapPin
+      position={position}
+      draggable={true}
+      ondragend={(lng: number) => {
+        position.lng = lng
+      }}
+    />
+  )
+}
