+++ packages/components/src/MapWithPin/MapWithPin.stories.tsx
@@ -1,17 +1,18 @@
-import { MapWithDraggablePin } from './MapWithDraggablePin'
+import { MapWithPin } from './MapWithPin'
 
 import type { Meta, StoryFn } from '@storybook/react'
 
 export default {
-  title: 'Components/MapWithDraggablePin',
-  component: MapWithDraggablePin,
-} as Meta<typeof MapWithDraggablePin>
+  title: 'Components/MapWithPin',
+  component: MapWithPin,
+} as Meta<typeof MapWithPin>
 
-export const Default: StoryFn<typeof MapWithDraggablePin> = () => {
+export const Default: StoryFn<typeof MapWithPin> = () => {
   const position = { lat: 0, lng: 0 }
   return (
-    <MapWithDraggablePin
+    <MapWithPin
       position={position}
+      draggable={true}
       updatePosition={(_position: { lat: number; lng: number }) => {
         position.lat = _position.lat
         position.lng = _position.lng
