+++ packages/components/src/index.ts
@@ -27,7 +27,7 @@ export { InternalLink } from './InternalLink/InternalLink'
 export { LinkifyText } from './LinkifyText/LinkifyText'
 export { Loader } from './Loader/Loader'
 export { MapMemberCard } from './MapMemberCard/MapMemberCard'
-export { MapWithDraggablePin } from './MapWithDraggablePin/MapWithDraggablePin'
+export { MapWithPin } from './MapWithPin/MapWithPin'
 export { MemberBadge } from './MemberBadge/MemberBadge'
 export { Modal } from './Modal/Modal'
 export { ModerationStatus } from './ModerationStatus/ModerationStatus'
