+++ src/pages/UserSettings/labels.ts
@@ -22,6 +22,7 @@ export const buttons = {
     type: 'type',
   },
   map: 'Add a map pin',
+  editPin: 'Edit map pin',
   removePin: 'Remove map pin',
   save: 'Save profile',
   success: 'Profile saved successfully',
@@ -140,7 +141,10 @@ export const headings = {
   map: {
     description:
       'Add yourself to the map as an individual who wants to get started. Find local community members and meetup to join forces and collaborate.',
-    title: 'Add yourself to the map!',
+    addPinTitle: 'Add yourself to the map 🌍',
+    yourPinTitle: 'Your map pin 🌍',
+    existingPinLabel:
+      'The map pin you registered has the following description:',
   },
   workspace: {
     description:
